renamed cloud outage simulation to cluster outage simulation
