+++ .github/workflows/contract-testing.yaml
@@ -19,7 +19,7 @@ jobs:
      access-list-api-key: ${{ steps.filter.outputs.access-list-api-key }}
      api-key: ${{ steps.filter.outputs.api-key }}
      privatelink-endpoint-service-data-federation-online-archive: ${{ steps.filter.outputs.privatelink-endpoint-service-data-federation-online-archive }}
-     cloud-outage-simulation: ${{ steps.filter.outputs.cloud-outage-simulation }}
+     cluster-outage-simulation: ${{ steps.filter.outputs.cluster-outage-simulation }}
      serverless-private-endpoint: ${{ steps.filter.outputs.serverless-private-endpoint }}
      data-lake-pipeline: ${{ steps.filter.outputs.data-lake-pipeline }}
      cloud-backup-restore-jobs: ${{ steps.filter.outputs.cloud-backup-restore-jobs }}
@@ -46,8 +46,8 @@ jobs:
           - 'cfn-resources/api-key/**'
         privatelink-endpoint-service-data-federation-online-archive:
           - 'cfn-resources/privatelink-endpoint-service-data-federation-online-archive/**'
-        cloud-outage-simulation:
-          - 'cfn-resources/cloud-outage-simulation/**'
+        cluster-outage-simulation:
+          - 'cfn-resources/cluster-outage-simulation/**'
         data-lake-pipeline:
           - 'cfn-resources/data-lake-pipeline/**'
         serverless-private-endpoint:
@@ -183,9 +183,9 @@ jobs:
        
        make run-contract-testing
        make delete-test-resources
-  cloud-outage-simulation :
+  cluster-outage-simulation :
     needs: change-detection
-    if: ${{ needs.change-detection.outputs.cloud-outage-simulation == 'true' }}
+    if: ${{ needs.change-detection.outputs.cluster-outage-simulation == 'true' }}
     runs-on: ubuntu-latest
     steps:
       - name: setup Atlas CLI
