+++ cfn-resources/cluster-outage-simulation/cmd/resource/model.go
