+++ cfn-resources/cluster-outage-simulation/cmd/resource/resource.go
