+++ cfn-resources/cluster-outage-simulation/docs/filter.md
@@ -1,4 +1,4 @@
-# MongoDB::Atlas::CloudOutageSimulation Filter
+# MongoDB::Atlas::ClusterOutageSimulation Filter
 
 ## Syntax
