+++ cfn-resources/cluster-outage-simulation/resource-role.yaml
