+++ cfn-resources/cluster-outage-simulation/template.yml
