+++ cfn-resources/cluster-outage-simulation/test/cfn-test-delete-inputs.sh
