+++ cfn-resources/cluster-outage-simulation/test/cluster-outage-simulation.sample-cfn-request.json
