+++ cfn-resources/cluster-outage-simulation/test/cluster.json
