+++ cfn-resources/cluster-outage-simulation/test/inputs_1_create.template.json
