+++ cfn-resources/cluster-outage-simulation/test/inputs_1_invalid.template.json
